+++ extension.js
@@ -1,83 +1,56 @@
 // The module 'vscode' contains the VS Code extensibility API
 // Import the module and reference it with the alias vscode in your code below
 const vscode = require('vscode');
-const ConfigurableCompletionItemOptions = require('./models/options/ConfigurableCompletionItemOptions');
-const ConfigurableDefinitionOptions = require('./models/options/ConfigurableDefinitionOptions');
-const ConfigurableDefinitionProvider = require('./models/providers/ConfigurableDefinitionProvider');
-const ConfigurableCompletionItemProvider = require('./models/providers/ConfigurableCompletionItemProvider');
+const providerRegistry = require('./models/services/ProviderRegistry');
 const Logger = require('./models/services/Logger');
 
 /**
  * @param {vscode.ExtensionContext} context
  */
 function activate(context) {
+	vscode.workspace.onDidChangeConfiguration(onConfigurationChanged);
+	providerRegistry.setContext(context);
+	registerProviders();
+}
+exports.activate = activate;
 
+/**
+ * Re-registers providers
+ * @param {vscode.ConfigurationChangeEvent} event 
+ */
+function onConfigurationChanged(event) {
+	if (!event.affectsConfiguration('htmlConfigurableAutocomplete')) {
+		return;
+	}
+	Logger.info(`Configuration updated, re-registering providers.`);
+	registerProviders();
+}
+
+function registerProviders() {
 	const configuration = vscode.workspace.getConfiguration('htmlConfigurableAutocomplete');
+
+	providerRegistry.clearAllProviders();
+
 	if (configuration.enable === false) {
+		Logger.warn(`The extension is disabled by configuration and won't do anything`);
 		return;
 	}
 	const completionItemOptions = configuration.completionItemProviders;
 	const definitionOptions = configuration.definitionProviders;
 
-	/**
-	 * @type {Array<vscode.Disposable>}
-	 */
-	const disposableList = [];
-
 	try {
-		const registeredCompletionItemProviders = registerCompletionItemProviders(completionItemOptions, disposableList);
-		Logger.info(`Registered ${registeredCompletionItemProviders} completion item providers`);
-		const registeredDefinitionProviders = registerDefinitionProviders(definitionOptions, disposableList);
-		Logger.info(`Registered ${registeredDefinitionProviders} definition providers`);
+		providerRegistry.registerCompletionItemProviders(completionItemOptions);
+		Logger.info(`Registered ${providerRegistry.completionItemProviders.length} completion item providers`);
+		providerRegistry.registerDefinitionProviders(definitionOptions);
+		Logger.info(`Registered ${providerRegistry.definitionProviders.length} definition providers`);
 	} catch (error) {
 		Logger.error(`${error}`);
 	}
 
-	if (disposableList.length == 0) {
+	if (providerRegistry.completionItemProviders.length + providerRegistry.definitionProviders.length == 0) {
 		Logger.warn(`Will not do anything since no (valid) rules for completionItemProviders and definitionProviders were set in the configuration.`);
 		return;
 	}
-	context.subscriptions.push(...disposableList);
-}
-exports.activate = activate;
-
-/**
- * @param {Array<any>} configurationRules
- * @param {Array<vscode.Disposable>} disposableList 
- */
-function registerCompletionItemProviders(configurationRules, disposableList) {
-	let counter = 0;
-
-	if (!configurationRules || !Array.isArray(configurationRules)) {
-		return counter;
-	}
-
-	configurationRules.forEach(configurationRule => {
-		const options = new ConfigurableCompletionItemOptions(configurationRule);
-		const disposable = vscode.languages.registerCompletionItemProvider({ scheme: 'file', language: 'html' }, new ConfigurableCompletionItemProvider(options), ...options.triggerCharacters);
-		disposableList.push(disposable);
-		counter++;
-	});
-	return counter;
-}
-
-/**
- * @param {Array<any>} configurationRules
- * @param {Array<vscode.Disposable>} disposableList 
- * @returns {number}
- */
-function registerDefinitionProviders(configurationRules, disposableList) {
-	let counter = 0;
-	if (!configurationRules || !Array.isArray(configurationRules)) {
-		return counter;
-	}
-	configurationRules.forEach(configurationRule => {
-		const options = new ConfigurableDefinitionOptions(configurationRule);
-		const disposable = vscode.languages.registerDefinitionProvider({ scheme: 'file', language: 'html'}, new  ConfigurableDefinitionProvider(options));
-		disposableList.push(disposable);
-		counter++;
-	});
-	return counter;
 }
 
 // this method is called when your extension is deactivated