+++ CHANGELOG.md
@@ -1,9 +1,18 @@
 # Change Log
 
-All notable changes to the "html-configurable-autocomplete" extension will be documented in this file.
-
-Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
 ## [Unreleased]
 
-- Initial release
+## [1.1.0] - 2017-07-28
+### Added
+ - Support for `${definition...}` placeholders in glob patterns for completion item providers.
+
+## [1.0.0] - 2017-07-26
+### Added
+ - Completion item providers.
+ - Definition providers.
+
+
+[Unreleased]: https://github.com/olivierlacan/keep-a-changelog/compare/v1.1.0...HEAD
+[1.1.0]: https://github.com/olivierlacan/keep-a-changelog/compare/v1.1.0...v1.0.0
+[1.0.0]: https://github.com/BrightSoul/vscode-html-configurable-autocomplete/releases/tag/v1.0.0